backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.2",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
[tool.setuptools]
packages = ["src"]
[tool.setuptools.package-data]
backend = ["*.py"]
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
main = "src.main:main"
dev = "src.main:dev"